Updated FCollada includes directory. Updated vcproj (added user32.lib to linker; renamed debug version to FColladaD.dll). Updated precompiled .libs.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6341 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/fcollada/include/LibXML/tree.h
@@ -1023,12 +1023,12 @@ XMLPUBFUN int XMLCALL
 XMLPUBFUN void XMLCALL		
 		xmlDocDumpFormatMemory	(xmlDocPtr cur,
 					 xmlChar **mem,
-					 int *size,
+					 intptr_t *size,
 					 int format);
 XMLPUBFUN void XMLCALL		
 		xmlDocDumpMemory	(xmlDocPtr cur,
 					 xmlChar **mem,
-					 int *size);
+					 intptr_t *size);
 XMLPUBFUN void XMLCALL		
 		xmlDocDumpMemoryEnc	(xmlDocPtr out_doc,
 					 xmlChar **doc_txt_ptr,
